+++ src/index.js
@@ -1,11 +1,5 @@
 import './style.css';
-import { addComment,updateUIWithComments,checkCommentsInput,updateUINumberOfComments, getComments,postComments} from './modules/comments';
-import { getLikes,
-  postLikes,
-  updateUI,
-
-  like,
-} from './modules/likes';
+import { getLikes, fetchAllMeals } from './modules/likes.js';
 
 const fontAwesome = document.createElement('link');
 fontAwesome.setAttribute('rel', 'stylesheet');
@@ -17,12 +11,35 @@ document.head.appendChild(fontAwesome);
 
 const mealsDiv = document.querySelector('.meals');
 const numberOfMeals = document.querySelector('.numberOfMeals');
-const fetchAllMeals = async () => {
-  const response = await fetch(
-    'https://www.themealdb.com/api/json/v1/1/search.php?s=',
-  );
-  const data = await response.json();
+
+const postLikes = async (id) => {
+  const uri = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/R0U3YhWaag3EdpAQTbkm/likes';
+  const response = await fetch(uri, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify({
+      item_id: id,
+    }),
+  });
+  const data = await response.text();
   return data;
+};
+
+const updateUI = (id, likes) => {
+  const meal = document.getElementById(id);
+  const mealInfoDiv = meal.querySelector('.mealInfo');
+  const mealLikesDiv = mealInfoDiv.querySelector('.likes');
+  mealLikesDiv.innerHTML = likes;
+};
+
+const like = async (id) => {
+  await postLikes(id);
+  getLikes().then((data) => {
+    const likesCount = data.filter((like) => like.item_id === id);
+    updateUI(id, likesCount[0]?.likes);
+  });
 };
 
 const fetchMealById = async (mealId) => {
@@ -44,7 +61,71 @@ const mealsWithLikes = async () => {
   return mealsWithLikes;
 };
 
-// popup
+const updateUIWithComments = (comment) => {
+  const commentsDiv = document.querySelector('.comments');
+  commentsDiv.innerHTML += `
+      <div class="comment">
+          <div class="comment-header">
+              <h5 class="comment-author">${comment.username}</h5>
+              <span class="comment-date">${comment.creation_date}</span>
+          </div>
+          <p class="comment-body">${comment.comment}</p>
+      </div>
+      `;
+};
+const checkCommentsInput = () => {
+  const username = document.getElementById('username');
+  const comment = document.getElementById('comment');
+  if (username.value === '' || comment.value === '') {
+    return false;
+  }
+  return true;
+};
+const postComments = async (id, comment, user) => {
+  const uri = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/R0U3YhWaag3EdpAQTbkm/comments';
+  const response = await fetch(uri, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify({
+      item_id: id,
+      username: user,
+      comment,
+    }),
+  });
+  const data = await response.text();
+  return data;
+};
+
+const getComments = async (id) => {
+  const uri = `https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/R0U3YhWaag3EdpAQTbkm/comments?item_id=${id}`;
+  const response = await fetch(uri);
+  const data = await response.json();
+  return data;
+};
+const updateUINumberOfComments = (comments) => {
+  const totalComments = document.querySelector('.comment-total');
+  totalComments.innerHTML = comments.length;
+};
+const clearCommentsInput = () => {
+  const username = document.getElementById('username');
+  const comment = document.getElementById('comment');
+  username.value = '';
+  comment.value = '';
+};
+const addComment = async (id, comment, user, e) => {
+  e.preventDefault();
+  if (!checkCommentsInput()) return;
+  const comments = await postComments(id, comment, user);
+  if (comments.error) return;
+  const Comments = await getComments(id);
+  const lastComment = Comments[Comments.length - 1];
+  updateUIWithComments(lastComment);
+  updateUINumberOfComments(Comments);
+  clearCommentsInput();
+};
+
 const popup = async (mealId) => {
   const meal = fetchMealById(mealId);
   const Comments = await getComments(mealId);
@@ -194,6 +275,8 @@ const popup = async (mealId) => {
     });
   });
 };
+
+// popup
 
 const listAllMeals = async () => {
   const allMeals = await mealsWithLikes();
